# swagger.yaml
swagger: "2.0"
info:
  description: "Simple shop system with basic functionality like products catalog / search / cart and order mechanism. Current version supports only small bunch of CRUD methods for user and product entities."
  version: "0.1.0"
  title: "Simple eCommerce System"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "igor@musitschenko.de"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
...
schemes:
  - "http"
paths:
...
  /products:
    get:
      tags:
        - "products"
      summary: "Returns all products"
      operationId: "getProducts"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
    post:
      tags:
        - "products"
      summary: "Add product to catalog"
      description: "This can only be done by the logged in user."
      operationId: "createProducts"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        200:
          description: "success"
        422:
          description: "Unprocessable Entity"
    put:
      tags:
        - "products"
      summary: "Update product data"
      description: "This can only be done by the logged in user."
      operationId: "updateProducts"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/ProductUpdate"
      responses:
        200:
          description: "success"
        422:
          description: "Unprocessable Entity"
definitions:
...
  Product:
    type: "object"
    required:
      - "sku"
      - "name"
      - "description"
    properties:
      sku:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
    xml:
      name: "Product"
  ProductUpdate:
    type: "object"
    required:
      - "id"
      - "sku"
      - "name"
      - "description"
    properties:
      id:
        type: "integer"
        format: "int64"
      sku:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
    xml:
      name: "Product"